requirements.txt
# FACT System Dependencies
# Core dependencies
anthropic>=0.25.6  # Latest version to fix proxy issues
groq>=0.4.0  # Groq API for fast inference
aiohttp==3.9.5
python-dotenv==1.0.1
pydantic==2.8.2
httpx>=0.24.0  # Required by newer anthropic versions

# Database
aiosqlite==0.20.0
asyncpg==0.29.0  # PostgreSQL async driver for Railway
psycopg2-binary==2.9.9  # PostgreSQL fallback driver

# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development dependencies
black==24.8.0
flake8==7.1.0
mypy==1.11.0

# Monitoring and logging
structlog==24.1.0

# Web server dependencies (for Railway deployment)
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.0.0
python-multipart==0.0.6  # Required for file upload endpoints

# Enhanced search dependencies
numpy==1.26.4  # For statistical calculations in enhanced search